+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### v0.20
 
+- Removed custom logger from ported code and added SLF4J.
 - Combined my_parser_wrapper.sh and tag.sh scripts.
 - Moved scripts from server-directory to scripts-directory.
 - Added parse_file.sh and parse_text.sh scripts.